action panel on craft window
